Moved exercise 2
