--- README.md
@@ -1,8 +1,0 @@
-# Links to exercises and solutions:
-
-
-## Exercise 2:
-exercise: https://mybinder.org/v2/gh/mie-lab/GDA20/master?filepath=gda_sdm_main.ipynb
-
-solution:
-
