+++ exercise2_spatialdatamining1/stops/stops.cpg
@@ -1,0 +1,1 @@
+ISO-8859-1
